Added User Auth & Restricted Routes

+++ base/views.py
@@ -1,7 +1,13 @@
 from django.shortcuts import redirect, render
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth import authenticate,login,logout
 from .models import Room, Topic
+from django.contrib.auth.models import User
 from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
+from django.http import HttpResponse
 # Create your views here.
 
 # rooms=[
@@ -9,6 +15,52 @@ from .forms import RoomForm
 #     {'id':2, 'name':'Design with!'},
 #     {'id':3, 'name':'Lets forget python!'},
 # ]
+
+def loginPage(request):
+    page= 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method=='POST':
+            username=request.POST.get('username').lower()
+            password=request.POST.get('password')
+
+            try:
+                user = User.objects.get(username=username)
+            except:
+                messages.error(request, "User doesn't exist")
+            
+            user = authenticate(request,username=username,password=password)
+
+            if user is not None:
+                login(request,user)
+                return redirect('home')
+            else:
+                messages.error(request,'Username or password does not exist')
+    context={'page':page}
+    return render(request,'base/login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page='register'
+    form=UserCreationForm()
+
+    if request.method == 'POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during registration')
+
+    context={'page':page,'form':form}
+    return render(request,'base/login_register.html',context)
 
 def home(request):
     q=request.GET.get('q') if request.GET.get('q') !=None else ''
@@ -30,6 +82,7 @@ def room(request,pk):
     context={'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method=='POST':
@@ -40,9 +93,13 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed to do this!!")
 
     if request.method=='POST':
         form = RoomForm(request.POST, instance=room)
@@ -53,8 +110,13 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room =Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed to do this!!")
+    
     if request.method=='POST':
         room.delete()
         return redirect('home')
